Avoid StackOverflowException when converting DTOs
The Event Converter Error Handler uses a recursive call to attempt a last-ditch conversion, unfortunately if this is triggered and fails then it causes the Error Handler to be called again with the same values. This cycles forever causing the problem.

The simple fix is to keep track of the object being converted and avoid trying to handle an error for it a second time